Fixing consumer bug - initalizer for conditional binding must have optional type not [String:SKPRoduct]
This error occurs because you are trying to access an instance member of self (storeManager) within the property initializer of a view. However, at the time the property initializer is executed, self (i.e., the view instance) is not fully initialized yet and the instance members are not available.

To fix the error, you can move the initialization of the BuySubscriptionOrItemsCard view to the onAppear closure of the view. This ensures that the storeManager instance has been fully initialized before accessing it. Here's an example: